+++ laktory/models/pipeline/_update_tables_metadata.py
@@ -23,7 +23,7 @@ def _update_tables_metadata():
     logger.info(f"Updating tables metadata for pipelines {filepaths}")
 
     for filepath in filepaths:
-        with open(filepath, "r") as fp:
+        with open(filepath, "r", encoding="utf-8-sig") as fp:
             if str(filepath).endswith(".yaml"):
                 pl = lk.models.Pipeline.model_validate_yaml(fp)
             else: